risolti problemi della precedente pull request

+++ src/main/java/com/develhope/spring/services/CommentServiceImpl.java
@@ -3,7 +3,7 @@ package com.develhope.spring.services;
 import com.develhope.spring.entities.Comment;
 import com.develhope.spring.repositories.CommentRepository;
 import org.springframework.beans.factory.annotation.Autowired;
-import java.util.List;
+import java.util.NoSuchElementException;
 import java.util.Optional;
 
 public class CommentServiceImpl implements CommentService{
@@ -12,23 +12,22 @@ public class CommentServiceImpl implements CommentService{
     CommentRepository commentRepository;
 
     @Override
-    public List<Comment> getAllComment() {
-        return commentRepository.findAll();
-    }
-
-    @Override
     public Comment getCommentById(Long id) {
         Optional<Comment> comment = commentRepository.findById(id);
-        return comment.orElse(null);
+        return comment.orElseThrow(()->new NoSuchElementException("Comment not found"));
     }
 
     @Override
     public Comment createComment(Comment comment) {
+        Optional<Comment> optionalComment = commentRepository.findById(comment.getId());
+        if(optionalComment.isPresent()){
+            throw new IllegalArgumentException("Comment already exist");
+        }
         return commentRepository.saveAndFlush(comment);
     }
 
     @Override
-    public Comment updateComment(Comment comment, Long id) throws Exception {
+    public Comment updateComment(Comment comment, Long id) {
         Optional <Comment> optionalComment= commentRepository.findById(id);
         if(optionalComment.isPresent()){
             Comment existComment = optionalComment.get();
@@ -36,17 +35,18 @@ public class CommentServiceImpl implements CommentService{
             existComment.setDateTime((comment.getDateTime()));
             return commentRepository.saveAndFlush(existComment);
         }else{
-            throw new Exception();
+            throw new NoSuchElementException("Comment not found");
         }
     }
 
     @Override
     public void deleteCommentById(Long id) {
-        commentRepository.deleteById(id);
-    }
+        Optional<Comment> optionalComment = commentRepository.findById(id);
+        if(optionalComment.isPresent()){
+            commentRepository.deleteById(id);
+        } else {
+            throw new NoSuchElementException("Comment not found");
+        }
 
-    @Override
-    public void deleteAllComment() {
-        commentRepository.deleteAll();
     }
 }
